frontend/requirements.txt
streamlit>=1.10
requests>=2.25
Pillow>=9.0
